profile-orchestrator/requirements.txt
requests
supabase
google-cloud-pubsub
google-genai
openai